+++ packages/kysely/migrate/migrations/2025-09-16T12:00:00_clients.ts
@@ -18,28 +18,18 @@ export async function up(db: Kysely<Database>): Promise<void> {
     .addColumn("oidc_conformant", "integer", (col) =>
       col.defaultTo(1).notNull(),
     )
-    .addColumn("callbacks", "text", (col) => col.defaultTo("[]").notNull())
-    .addColumn("allowed_origins", "text", (col) =>
-      col.defaultTo("[]").notNull(),
-    )
-    .addColumn("web_origins", "text", (col) => col.defaultTo("[]").notNull())
-    .addColumn("client_aliases", "text", (col) => col.defaultTo("[]").notNull())
-    .addColumn("allowed_clients", "text", (col) =>
-      col.defaultTo("[]").notNull(),
-    )
-    .addColumn("allowed_logout_urls", "text", (col) =>
-      col.defaultTo("[]").notNull(),
-    )
-    .addColumn("session_transfer", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
-    .addColumn("oidc_logout", "text", (col) => col.defaultTo("{}").notNull())
-    .addColumn("grant_types", "text", (col) => col.defaultTo("[]").notNull())
-    .addColumn("jwt_configuration", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
-    .addColumn("signing_keys", "text", (col) => col.defaultTo("[]").notNull())
-    .addColumn("encryption_key", "text", (col) => col.defaultTo("{}").notNull())
+    .addColumn("callbacks", "text", (col) => col.notNull())
+    .addColumn("allowed_origins", "text", (col) => col.notNull())
+    .addColumn("web_origins", "text", (col) => col.notNull())
+    .addColumn("client_aliases", "text", (col) => col.notNull())
+    .addColumn("allowed_clients", "text", (col) => col.notNull())
+    .addColumn("allowed_logout_urls", "text", (col) => col.notNull())
+    .addColumn("session_transfer", "text", (col) => col.notNull())
+    .addColumn("oidc_logout", "text", (col) => col.notNull())
+    .addColumn("grant_types", "text", (col) => col.notNull())
+    .addColumn("jwt_configuration", "text", (col) => col.notNull())
+    .addColumn("signing_keys", "text", (col) => col.notNull())
+    .addColumn("encryption_key", "text", (col) => col.notNull())
     .addColumn("sso", "integer", (col) => col.defaultTo(0).notNull())
     .addColumn("sso_disabled", "integer", (col) => col.defaultTo(1).notNull())
     .addColumn("cross_origin_authentication", "integer", (col) =>
@@ -52,43 +42,33 @@ export async function up(db: Kysely<Database>): Promise<void> {
     .addColumn("custom_login_page", "text")
     .addColumn("custom_login_page_preview", "text")
     .addColumn("form_template", "text")
-    .addColumn("addons", "text", (col) => col.defaultTo("{}").notNull())
+    .addColumn("addons", "text", (col) => col.notNull())
     .addColumn("token_endpoint_auth_method", "varchar(64)", (col) =>
       col.defaultTo("client_secret_basic"),
     )
-    .addColumn("client_metadata", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
-    .addColumn("mobile", "text", (col) => col.defaultTo("{}").notNull())
+    .addColumn("client_metadata", "text", (col) => col.notNull())
+    .addColumn("mobile", "text", (col) => col.notNull())
     .addColumn("initiate_login_uri", "varchar(2083)")
-    .addColumn("native_social_login", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
-    .addColumn("refresh_token", "text", (col) => col.defaultTo("{}").notNull())
-    .addColumn("default_organization", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
+    .addColumn("native_social_login", "text", (col) => col.notNull())
+    .addColumn("refresh_token", "text", (col) => col.notNull())
+    .addColumn("default_organization", "text", (col) => col.notNull())
     .addColumn("organization_usage", "varchar(32)", (col) =>
       col.defaultTo("deny"),
     )
     .addColumn("organization_require_behavior", "varchar(32)", (col) =>
       col.defaultTo("no_prompt"),
     )
-    .addColumn("client_authentication_methods", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
+    .addColumn("client_authentication_methods", "text", (col) => col.notNull())
     .addColumn("require_pushed_authorization_requests", "integer", (col) =>
       col.defaultTo(0).notNull(),
     )
     .addColumn("require_proof_of_possession", "integer", (col) =>
       col.defaultTo(0).notNull(),
     )
-    .addColumn("signed_request_object", "text", (col) =>
-      col.defaultTo("{}").notNull(),
-    )
+    .addColumn("signed_request_object", "text", (col) => col.notNull())
     .addColumn("compliance_level", "varchar(64)")
     .addColumn("par_request_expiry", "integer")
-    .addColumn("token_quota", "text", (col) => col.defaultTo("{}").notNull())
+    .addColumn("token_quota", "text", (col) => col.notNull())
     .addColumn("created_at", "varchar(35)", (col) => col.notNull())
     .addColumn("updated_at", "varchar(35)", (col) => col.notNull())
     .addPrimaryKeyConstraint("clients_tenant_id_client_id", [