+++ packages/kysely/test/sessions/get.spec.ts
@@ -17,14 +17,16 @@ describe("sessions", () => {
       });
 
       // Add a client
-      await data.applications.create("tenantId", {
-        id: "clientId",
+      await data.clients.create("tenantId", {
+        client_id: "client123",
         client_secret: "clientSecret",
         name: "Test Client",
         callbacks: ["https://example.com/callback"],
         allowed_logout_urls: ["https://example.com/callback"],
         web_origins: ["https://example.com"],
-        disable_sign_ups: false,
+        client_metadata: {
+          disable_sign_ups: "false",
+        },
       });
 
       // Add a test user